Native code refractor
