+++ android/app/build.gradle
@@ -73,7 +73,7 @@ import com.android.build.OutputFile
  */
 
 project.ext.react = [
-    entryFile: "index.js"
+    entryFile: "index.android.js"
 ]
 
 apply from: "../../node_modules/react-native/react.gradle"
