+++ app/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Provider } from 'react-redux';
-import store from './store';
+import store from '../src/store/index';
 import AppRouter from './router';
