+++ app/base_components/PrimaryText.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import Color from '../constants/colors';
+import Color from '../../src/constants/colors';
 
 const PrimaryText = styled.Text`
   width: 100%;
