+++ app/base_components/RippleIcon.js
