+++ app/base_components/Section.js
