+++ app/base_components/ViewRow.js
