+++ app/components/Checkout/Footer.js
