+++ app/components/RestaurantItem.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { Image, TouchableOpacity, View } from 'react-native';
 import PropTypes from 'prop-types';
 
-import Assets from '../constants/assets';
+import Assets from '../../src/constants/assets';
 import PrimaryText from '../base_components/PrimaryText';
 import SecondaryText from '../base_components/SecondaryText';
