+++ index.android.js
@@ -1,4 +1,4 @@
 import { AppRegistry } from 'react-native';
-import App from './src/App';
+import App from './app/App';
 
 AppRegistry.registerComponent('RestaurantApp', () => App);
