+++ index.ios.js
@@ -1,0 +1,4 @@
+import { AppRegistry } from 'react-native';
+import App from './app/App';
+
+AppRegistry.registerComponent('RestaurantApp', () => App);
